Add: 브랜드명 합치기
