+++ src/index.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import Main from './pages/Main/Main';
-import Test from './pages/Test/Test';
+// import Test from './pages/Test/Test';
 ReactDOM.render(
   <React.StrictMode>
-    {/* <Main /> */}
-    <Test />
+    <Main />
+    {/* <Test /> */}
   </React.StrictMode>,
   document.getElementById('root')
 );
